Lab 3 completed

+++ lab3/Flik/HorribleSteve.java
@@ -1,7 +1,11 @@
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
 public class HorribleSteve {
     public static void main(String [] args) throws Exception {
-        int i = 0;
-        for (int j = 0; i < 500; ++i, ++j) {
+        double i = 0;
+        for (double j = 0; i < 500; ++i, ++j) {
             if (!Flik.isSameNumber(i, j)) {
                 throw new Exception(
                         String.format("i:%d not same as j:%d ??", i, j));
